refactor - removing the auth submodule to merge it into proms_mcp - removing teh clusterrolebinding which is now useless
