+++ openshift/deploy.yaml
@@ -57,21 +57,6 @@ objects:
     labels:
       app: proms-mcp-server
 
-- apiVersion: rbac.authorization.k8s.io/v1
-  kind: ClusterRoleBinding
-  metadata:
-    name: proms-mcp-auth-delegator
-    labels:
-      app: proms-mcp-server
-  subjects:
-  - kind: ServiceAccount
-    name: proms-mcp-server
-    namespace: ${NAMESPACE}
-  roleRef:
-    kind: ClusterRole
-    name: system:auth-delegator
-    apiGroup: rbac.authorization.k8s.io
-
 - apiVersion: apps/v1
   kind: Deployment
   metadata:
